requirements.txt
flask
torch
timm
torchvision
Pillow
gunicorn
requests
